[ML] Move classification and regression loss functions into their own file (#1027)
