+++ lib/api/CDataFrameTrainBoostedTreeClassifierRunner.cc
@@ -12,6 +12,7 @@
 
 #include <maths/CBoostedTree.h>
 #include <maths/CBoostedTreeFactory.h>
+#include <maths/CBoostedTreeLoss.h>
 #include <maths/CDataFramePredictiveModel.h>
 #include <maths/CDataFrameUtils.h>
 #include <maths/COrderings.h>
