+++ lib/api/unittest/CDataFrameAnalyzerTrainingTest.cc
@@ -13,6 +13,7 @@
 
 #include <maths/CBoostedTree.h>
 #include <maths/CBoostedTreeFactory.h>
+#include <maths/CBoostedTreeLoss.h>
 #include <maths/CTools.h>
 
 #include <api/CDataFrameAnalyzer.h>
