+++ lib/maths/unittest/CBoostedTreeTest.cc
@@ -12,6 +12,7 @@
 #include <maths/CBasicStatistics.h>
 #include <maths/CBoostedTree.h>
 #include <maths/CBoostedTreeFactory.h>
+#include <maths/CBoostedTreeLoss.h>
 #include <maths/CSolvers.h>
 #include <maths/CTools.h>
